fastapi-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
mysql-connector-python==9.2.0
python-dotenv==1.0.1
python-multipart==0.0.20
